docs: structure for the "Research Purpose" section of the proposal
esoteric question: when the entire project is a written document, what is a doc, and what is a feat? I suppose that once I start making experiments _those_ won't be docs, but it's starting to feel bizarre to refer to every portion of this phase as a doc (proposal, journal, etc) Conversations.md is being committed with the tag 'conversations' so maybe doc should just be used or the journaly bits? But also, the Proposal is legitimately a document? Anywho